web-soap-service/requirements.txt
spyne
flask
mysql-connector-python
python-dotenv
lxml 